requirements.txt
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# LangChain / LangGraph
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
litellm>=1.20.0

# Anthropic SDK (for fusion agent LLM calls)
anthropic>=0.39.0

# Video Processing
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# Chess
python-chess>=1.999

# Audio/Transcript
openai-whisper>=20231117
pydub>=0.25.1

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.2.1

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.11.0
ruff>=0.1.6
